extracted the InclusiveInterval into a separate module
